+++ doc/source/changelog.rst
@@ -5,6 +5,8 @@ v0.0.3
 ^^^^^^
 
 - Pages can be searched from the right top search box.
+- Show sub-pages when the accessed page does not exist but at least
+  one sub-page exists.
 - Use reversed :term:`page path` as a page title.
 
 v0.0.2